refactor: Moved lib/FileType to brim/ingest
