+++ src/js/brim/ingest/detectFileTypes.test.js
@@ -6,16 +6,18 @@ import ingest from "./"
 
 const testFile = (name) => path.join(__dirname, "../../../../itest/", name)
 const pcap = testFile("sample.pcap")
-const zeek = testFile("sample.zeek.log")
-const json = testFile("sample.json.zeek.log")
+const pcapng = testFile("sample.pcapng")
+const zeek = testFile("sample.tsv")
+const json = testFile("sample.ndjson")
 const unknown = testFile("setup.js")
 
 test("add file types", async () => {
-  let paths = [pcap, zeek, json, unknown]
+  let paths = [pcap, pcapng, zeek, json, unknown]
   let types = await ingest.detectFileTypes(paths)
 
   expect(types).toEqual([
     {type: "pcap", path: pcap},
+    {type: "pcap", path: pcapng},
     {type: "zeek", path: zeek},
     {type: "json", path: json},
     {type: "unknown", path: unknown}
